mscv warning clean

+++ test/di_mt_03.cpp
@@ -66,7 +66,7 @@ struct c3
 
 BOOST_AUTO_TEST_CASE(create_complex) {
     const int i = 42;
-    const int d = 42.0;
+    const double d = 42.0;
 
     typedef di::injector<
         di::policy<
